requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# Email
email-validator==2.2.0


# Utilities
python-dateutil==2.9.0.post0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1